+++ R/data.R
@@ -3,7 +3,7 @@
 #' Water quality data from the California Department of Fish and Wildlife 20mm survey.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 10,098 rows and 13 variables
+#' @format a tibble with 10,469 rows and 14 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
@@ -15,8 +15,9 @@
 #'   \item{Depth}{Bottom depth (m).}
 #'   \item{Tide}{Tidal stage.}
 #'   \item{Secchi}{Secchi depth (cm).}
-#'   \item{Temperature}{Temperature in °C.}
+#'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{Conductivity_bottom}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at bottom.}
 #'   \item{Notes}{Comments.}
 #'   }
 #' @details More metadata and information on methods are available \href{https://wildlife.ca.gov/Conservation/Delta/20mm-Survey}{here}.
@@ -28,12 +29,13 @@
 #' Water quality data from the California Department of Fish and Wildlife Bay Study.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 21,264 rows and 12 variables
+#' @format a tibble with 21,836 rows and 14 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
 #'   \item{Latitude}{Latitude in decimal degrees.}
 #'   \item{Longitude}{Longitude in decimal degrees.}
+#'   \item{Field_coords}{Were lat/long coordinates collected in the field (TRUE), or do they represent the location of a fixed station (FALSE)?}
 #'   \item{Date}{Date sample was collected.}
 #'   \item{Datetime}{Date and time of sample collection.}
 #'   \item{Depth}{Bottom depth (m).}
@@ -42,6 +44,7 @@
 #'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Temperature_bottom}{Temperature (°C) at bottom.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{Conductivity_bottom}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at bottom.}
 #'   }
 #' @details More metadata and information on methods are available \href{http://www.dfg.ca.gov/delta/projects.asp?ProjectID=BAYSTUDY}{here}.
 #' @seealso \code{\link{wq}}
@@ -67,6 +70,8 @@
 #'   \item{Temperature}{Temperature in °C.}
 #'   \item{Temperature_bottom}{Temperature (°C) at bottom.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
+#'   \item{DissolvedOxygen_bottom}{Dissolved oxygen (mg/L) at bottom.}
 #'   \item{Notes}{Comments.}
 #'   }
 #' @details More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.415.1}{here}.
@@ -161,6 +166,7 @@
 #'   \item{Secchi}{Secchi depth (cm).}
 #'   \item{Temperature}{Temperature in °C.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
 #'   }
 #' @details More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.244.3}{here}.
 #' @seealso \code{\link{wq}}
@@ -197,7 +203,7 @@
 #' Water quality data from the California Department of Water Resources Stockton Dissolved Oxygen monitoring.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 3,108 rows and 11 variables
+#' @format a tibble with 3,112 rows and 16 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
@@ -210,6 +216,11 @@
 #'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Temperature_bottom}{Temperature (°C) at bottom.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{Conductivity_bottom}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at bottom.}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
+#'   \item{DissolvedOxygen_bottom}{Dissolved oxygen (mg/L) at bottom.}
+#'   \item{pH}{pH (dimensionless) at surface.}
+#'   \item{pH_bottom}{pH (dimensionless) at bottom.}
 #'   }
 #' @details More metadata and information on methods are available \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.276.2}{here}.
 #' @seealso \code{\link{wq}}
@@ -220,7 +231,7 @@
 #' Water quality data from the California Department of Fish and Wildlife Spring Kodiak Trawl.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 4,309 rows and 13 variables
+#' @format a tibble with 4,505 rows and 13 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
@@ -308,6 +319,8 @@
 #'   \item{Secchi}{Secchi depth (cm).}
 #'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
+#'   \item{DissolvedOxygenPercent}{Dissolved oxygen percent (dimensionless) at surface.}
 #'   }
 #' @details More metadata and information on methods are available \href{https://watershed.ucdavis.edu/project/suisun-marsh-fish-study}{here}.
 #' @seealso \code{\link{wq}}
@@ -342,21 +355,32 @@
 #' Water quality data from the United States Geological Survey San Francisco Bay Water Quality Survey.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 23,021 rows and 17 variables
+#' @format a tibble with 23,923 rows and 22 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
 #'   \item{Latitude}{Latitude in decimal degrees.}
 #'   \item{Longitude}{Longitude in decimal degrees.}
 #'   \item{Date}{Date sample was collected.}
-#'   \item{Datetime}{Date and time of sample collection.}
-#'   \item{Sample_depth_surface}{Depth (m) of surface sample.}
-#'   \item{Sample_depth_bottom}{Depth (m) of bottom sample.}
+#'   \item{Datetime}{Date and time of sample collection. Reported as an average
+#'     when the collection times varied among the the surface and bottom WQ and
+#'     nutrient parameters.}
+#'   \item{Sample_depth_surface}{Depth (m) of surface sample. Reported as an
+#'     average when surface depths varied among the WQ parameters.}
+#'   \item{Sample_depth_bottom}{Depth (m) of bottom sample. Reported as an
+#'     average when bottom depths varied among the WQ parameters.}
 #'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Temperature_bottom}{Temperature (°C) at bottom.}
 #'   \item{Salinity}{Salinity at surface.}
+#'   \item{Salinity_bottom}{Salinity at bottom.}
 #'   \item{Chlorophyll}{Chlorophyll concentration (\eqn{\mu}g \ifelse{html}{\out{L<sup>-1</sup>}}{\eqn{L^{-1}}}) at surface.}
-#'   \item{Sample_depth_nutr_surface}{Depth (m) paired w/ nutrient sampling (range: 0-5 m).}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
+#'   \item{DissolvedOxygen_bottom}{Dissolved oxygen (mg/L) at bottom.}
+#'   \item{DissolvedOxygenPercent}{Dissolved oxygen percent (dimensionless) at surface.}
+#'   \item{DissolvedOxygenPercent_bottom}{Dissolved oxygen percent (dimensionless) at bottom.}
+#'   \item{Sample_depth_nutr_surface}{Depth (m) paired w/ nutrient sampling
+#'     (range: 0-5 m). Reported as an average when surface depths varied among the
+#'     nutrient parameters.}
 #'   \item{DissNitrateNitrite}{Dissolved Nitrate and Nitrite (mg/L).}
 #'   \item{DissAmmonia}{Dissolved Ammonia (mg/L).}
 #'   \item{DissOrthophos}{Dissolved Ortho-phosphate (mg/L).}
@@ -371,7 +395,7 @@
 #' Discrete water quality data from the USGS California Water Science Center
 #'
 #' @encoding UTF-8
-#' @format a tibble with 3,597 rows and 15 variables
+#' @format a tibble with 16,751 rows and 19 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
@@ -388,6 +412,10 @@
 #'   \item{DOC}{Dissolved Organic Carbon (mg/L)}
 #'   \item{DissOrthophos_Sign}{Whether the Dissolved Orthophosphate value is lower than reported ("<" because it is below the reporting limit and the reporting limit is used as the value), estimated "~", or reported as the measured value "=".}
 #'   \item{DissOrthophos}{Dissolved Ortho-phosphate (mg/L)}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
+#'   \item{pH}{pH (dimensionless) at surface.}
+#'   \item{Temperature}{Temperature (°C) at surface.}
+#'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
 #'   }
 #' @details More metadata and information on methods are available \href{https://nwis.waterdata.usgs.gov/usa/nwis/qwdata}{here for data} and \href{https://help.waterdata.usgs.gov/codes-and-parameters}{here for metadata}.
 #' @seealso \code{\link{wq}}
@@ -398,7 +426,7 @@
 #' Water quality data from the California Department of Water Resources Yolo Bypass Fish Monitoring Program.
 #'
 #' @encoding UTF-8
-#' @format a tibble with 8,874 rows and 12 variables
+#' @format a tibble with 8,883 rows and 14 variables
 #' \describe{
 #'   \item{Source}{Name of the source dataset.}
 #'   \item{Station}{Station where sample was collected.}
@@ -411,6 +439,8 @@
 #'   \item{Secchi}{Secchi depth (cm).}
 #'   \item{Temperature}{Temperature (°C) at surface.}
 #'   \item{Conductivity}{Specific conductance (\eqn{\mu}S \ifelse{html}{\out{cm<sup>-1</sup>}}{\eqn{cm^{-1}}}) at surface.}
+#'   \item{DissolvedOxygen}{Dissolved oxygen (mg/L) at surface.}
+#'   \item{pH}{pH (dimensionless) at surface.}
 #'   \item{Notes}{Notes or comments.}
 #'   }
 #'