+++ data-raw/SKT.R
@@ -39,7 +39,11 @@ SKT <- read_csv(file.path("data-raw", "SKT", "tblSample.csv"),
   separate(Longitude, into=c("LonD", "LonM", "LonS"), sep="-", remove=TRUE, convert=TRUE)%>%
   mutate(Latitude=LatD+LatM/60+LatS/3600,
          Longitude=(LonD+LonM/60+LonS/3600)*-1)%>%
-  mutate(Datetime = parse_date_time(paste0(Date, " ", hour(Time), ":", minute(Time)), "%Y-%m-%d %H:%M", tz="America/Los_Angeles"))%>%
+  mutate(
+    # Convert one time stamp from midnight to noon
+    Time = if_else(Date == "2018-03-06" & Station == "519", ymd_hms("1899-12-30 12:00:00", tz = "America/Los_Angeles"), Time),
+    Datetime = parse_date_time(paste0(Date, " ", hour(Time), ":", minute(Time)), "%Y-%m-%d %H:%M", tz="America/Los_Angeles")
+  )%>%
   select(-Time, -LatD, -LatM, -LatS, -LonD, -LonM, -LonS)%>%
   mutate(Tide=recode(as.character(Tide), `4`="Flood", `3`="Low Slack", `2`="Ebb", `1`="High Slack", `0`=NA_character_),
          Depth = Depth*0.3048)%>% # Convert feet to meters