+++ NEWS.md
@@ -1,8 +1,13 @@
 # discretewq (development version)
 
-* Updated to newest EMP data againwith some historical data issues fixed. 
+* Removed `pH_Sign` and `Turbidity_Sign` variables from `NCRO` because they were not needed. 
+* Added temperature and conductivity to `USGS_CAWSC`
+* Added DO and pH data to all surveys that collect this data. `USGS_SFBS` collects both calculated (from a sensor) and discrete DO, so we used discrete DO up to 2016 and calculated DO afterwards to mirror the methodological change that occurred in the `EMP` survey in 2016. 
+* Updated to newest EMP data with some historical data issues fixed. 
 * Added NCRO chlorophyll data
-* Updated STN, FMWT, EDSM, DJFMP, SLS, Suisun, EMP, USGS_SFBS, USGS_CAWSC, YBFMP, and 20mm datasets
+* Added bottom conductivity to `twentymm`, `baystudy`, and `SDO`
+* Fixed timezones for `SDO` data. SDO times are reported in PST but had incorrectly been imported as local time (PST/PDT). Now, they are imported as Etc/GMT+8 and then converted to America/Los_Angeles to correspond to the other surveys.
+* Updated STN, FMWT, EDSM, DJFMP, SLS, Suisun, EMP, USGS_SFBS, USGS_CAWSC, YBFMP, SKT, 20mm, baystudy, and SDO datasets
 
 # discretewq 2.3.2